+++ react-components/src/components/ProductFinder/ProductFinderModal.jsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from 'react'
 import PropTypes from 'prop-types'
 import ReactModal from 'react-modal'
 import Services from '@src/Services'
+import actions from '@src/actions'
 import { capitalize } from '@src/Helpers'
 import Spinner from '../Spinner/Spinner'
 import Interaction from './Interaction'
@@ -13,10 +14,10 @@ import { analytics } from '../../Helpers'
 
 
 export default function ProductFinderModal(props) {
-  const { modalIsOpen, setIsOpen, setSelectedProduct } = props;
+  const { modalIsOpen, setIsOpen, selectedProduct} = props;
 
   let scrollOuter
-
+  const [isSearching, setSearching] = useState(false)
   const [searchResults, setSearchResults] = useState()
   const [isLoading, setLoading] = useState(false)
   const [isScrolled, setIsScrolled] = useState(false)
@@ -37,6 +38,7 @@ export default function ProductFinderModal(props) {
 
   const closeModal = () => {
     setIsOpen(false)
+    setSearching(false)
     setSearchResults()
   }
 
@@ -65,30 +67,18 @@ export default function ProductFinderModal(props) {
   }
 
   const saveProduct = (commodityCode, commodityName) => {
-
-    setSelectedProduct({
-      name: commodityName,
-      code: commodityCode
-    })
-    analytics({
-      event: 'addProductSuccess',
-      productKeyword: capitalize(searchResults.productDescription),
-      productCode: commodityCode
-    })
-
-    Services.updateExportPlan({
-        export_commodity_codes: [{
-          commodity_name: commodityName,
-          commodity_code: commodityCode
-        }]
+    Services.store.dispatch(actions.setProduct({
+      commodity_name: commodityName,
+      commodity_code: commodityCode
+    }))
+    if (searchResults) {
+      analytics({
+        event: 'addProductSuccess',
+        productKeyword: capitalize(searchResults.productDescription),
+        productCode: commodityCode
       })
-      .then(() => {
-        closeModal()
-        window.location.reload()
-      })
-      .catch(() => {
-        closeModal()
-      })
+    }
+    closeModal()
   }
 
   const renderSearchResults = (newSearchResults) => {
@@ -158,13 +148,14 @@ export default function ProductFinderModal(props) {
   }
 
   const backToSearch = () => {
+    setSearching(true)
     renderSearchResults()
   }
 
   const Section = (title, sectionDetails) => {
     if (!sectionDetails || sectionDetails.length === 0 || !sectionDetails.map) return ''
     return (
-      <section className="p-h-l">
+      <section className="p-h-s">
         <h3 className="h-m p-v-xs">{title}</h3>
           {(sectionDetails || []).map((value) => {
             return value.type === 'SELECTION' ? 
@@ -210,7 +201,7 @@ export default function ProductFinderModal(props) {
 
   const sectionAssumptions = (sectionDetails) => {
     return sectionDetails && sectionDetails.length && (
-      <section className="p-h-l">
+      <section className="p-h-s">
         <h3 className="h-s p-0">Assumptions</h3>
         <p className="m-v-xxs">We&apos;ve answered some questions for you. View and change these if they&apos;re wrong.</p>
         <ExpandCollapse buttonLabel={ `View assumptions (${sectionDetails.length})`} expandedButtonLabel="Hide assumptions">{readOnlyContent(sectionDetails)}</ExpandCollapse> 
@@ -221,19 +212,19 @@ export default function ProductFinderModal(props) {
   const sectionProductDetails = (sectionDetails) => {
     return sectionDetails && sectionDetails.length && (
       <div>
-        <section className="p-h-l">
+        <section className="p-h-s">
           <h3 className="h-s p-0">Product details</h3>
           <p className="m-v-xxs">Things you&apos;ve told us about your product.</p>
           {readOnlyContent(sectionDetails)}
         </section>
-        <hr className="hr bg-black-10 m-h-l" />
+        <hr className="hr bg-black-10 m-h-s" />
       </div>
     ) || ''
   }
 
   const sectionFound = (_searchResults) => {
     return (
-      <section className="m-h-l m-b-s body-l">
+      <section className="m-h-s m-b-s body-l">
         <div className="h-m p-b-s">Match found</div>
         <StartEndPage 
           commodityCode={_searchResults.hsCode} 
@@ -246,7 +237,7 @@ export default function ProductFinderModal(props) {
 
   const sectionNoResults = (_searchResults) => {
     return (
-      <section className="m-h-l">
+      <section className="m-h-s">
         <div className="box box--no-pointer p-h-s p-t-s p-b-xs m-t-xs">
           <h3 className="h-m p-t-xxs">No results found for &lsquo;{_searchResults.productDescription}&rsquo;</h3>
           <h4 className="h-xs">Search tips</h4>
@@ -267,7 +258,7 @@ export default function ProductFinderModal(props) {
   }
   const sectionMultiItem = () => {
     return (
-      <section className="m-h-l">
+      <section className="m-h-s">
         <div className="box box--no-pointer p-h-s p-v-xs m-t-xs">
           <p>
             The item you are classifying is considered a complex item (or set) which normally requires each component to be classified separately. 
@@ -331,7 +322,7 @@ export default function ProductFinderModal(props) {
       <div>
         {!showingInteraction && searchResults.hsCode && sectionFound(searchResults)}
         {(!searchResults.hsCode || showingInteraction) && Section(`Tell us more about "${searchResults.currentItemName}"`, questions)}
-        {((known || questions) && !showingInteraction ) ? (<hr className="hr bg-red-deep-100 m-h-l"/>) : ''}
+        {((known || questions) && !showingInteraction ) ? (<hr className="hr bg-red-deep-100 m-h-s"/>) : ''}
         {!showingInteraction && sectionProductDetails(known)} 
         {!showingInteraction && sectionAssumptions(assumptions)}
       </div>
@@ -341,9 +332,9 @@ export default function ProductFinderModal(props) {
 
   const searchBox = () => {
     return (
-      <div className="p-h-l p-v-l">
-        <h3 className="h-m p-t-0 p-b-xxs">Search by name</h3>
-        <div>Find the product you want to export</div>
+      <div className="p-h-s p-t-l">
+        <h3 className="h-m p-t-0 p-b-xxs">Add product</h3>
+        <div>Adding a product personalises lessons and other content for you.</div>
         <div className="flex-centre m-t-xs search-input">
             <SearchInput
               id="search-input"
@@ -361,10 +352,76 @@ export default function ProductFinderModal(props) {
             <i className="fa fa-arrow-right"/>
           </button>
         </div>
+        <div className="box box--no-pointer m-t-s">
+          When you search for a product you may have to answer a few questions before you find a match.
+        </div>
+        <div className="box box--no-pointer m-t-s">
+          <p className="m-t-0 m-b-xs">This is because we use HS (or <span className="body-l-b">harmonised system</span>) codes to classify goods.</p>
+          <p className="m-v-0">Think of it like the folder structure on a computer.</p>
+        </div>
+        <div className="box box--no-pointer m-t-s m-h-0 grid">
+          <div className="c-1-2 p-h-0">
+            <p className="m-t-0 m-b-xs">You might see your product as "delicious green apples from the valley".</p>
+            <p className="m-v-0">But the system sees 'fruits; apples; fresh".</p>
+          </div>
+          <div className="c-1-2">
+            <img className="w-full" src="/static/images/apples-oranges-with-hs6.svg " alt="" />
+          </div>
+        </div>
+        <div className="box box--no-pointer m-t-s">
+          <p className="m-t-0 m-b-xs">You don't have to find a perfect match.</p>
+          <p className="m-v-0">Find a close match, then feel free to relabel it.</p>
+        </div>
       </div>
     )
   }
 
+  const showProduct = () => {
+    // When modal is opened - it shows the last selected product 
+    // rather than jumping directly into search
+    return (
+      <div className="p-h-s p-t-l">
+        <section className="m-b-s body-l">
+          <h2 className="h-m p-b-s">Your product</h2>
+          <StartEndPage 
+            commodityCode={selectedProduct.commodity_code || ''} 
+            defaultCommodityName={selectedProduct.commodity_name || ''} 
+            saveProduct={saveProduct}
+            label="You can rename this to match you exact product"
+            buttonLabel="Update product name"
+          />
+          <p>If you&apos;ve created an Export Plan, make sure you update it to reflect your new product. You can change product at any time.</p>
+          <button 
+            type="button" 
+            className="button button--tertiary button--icon" 
+            onClick={backToSearch} 
+          >
+            <i className="fas fa-arrow-left" />
+            Search again
+          </button>
+        </section>
+      </div>
+    )
+  }
+
+  const searchPages = () => {
+    // When in searching mode.  If there are searchResults will show a refinement/result page
+    // otherwise the search box page.
+    return !searchResults ? searchBox() : (
+      <span>
+        <button 
+          type="button" 
+          className="back-button m-f-s m-t-m" 
+          onClick={backToSearch} 
+        >
+          <i className="fa fa-arrow-circle-left m-r-xs"/>
+          Search again
+        </button>
+      {searchResults && resultsDisplay(searchResults)}
+      </span>
+    )
+  }
+  
   const scrollerClass = `scroll-area ${isScrolled && isScrolled.top ? 'scroll-shadow-top' : ''} ${isScrolled && isScrolled.bottom ? 'scroll-shadow-bottom' : ''}`
   const headerHeight = '0px'
 
@@ -379,9 +436,9 @@ export default function ProductFinderModal(props) {
         shouldCloseOnOverlayClick={false}
         style={{
           content:{
-            width:'auto',
-            left: '100px',
-            right: '100px',
+            maxWidth:'630px',
+            left: 'auto',
+            right: 'auto',
             overflow: 'hidden',
           }
         }}
@@ -407,16 +464,7 @@ export default function ProductFinderModal(props) {
               className="scroll-inner p-b-m"
               ref={(_scrollInner) => {scrollOuter = _scrollInner || scrollOuter}}
             >
-              {!searchResults ? searchBox() : (
-                <button 
-                  type="button" 
-                  className="back-button m-f-l m-t-m" 
-                  onClick={backToSearch} 
-                >
-                  <i className="fa fa-arrow-circle-left m-r-xs"/>
-                  Search again
-                </button>)}
-              {searchResults && resultsDisplay(searchResults)}
+              {(isSearching || !selectedProduct) ? searchPages() : showProduct()}
             </div>
           </div>
         </form>
@@ -428,5 +476,11 @@ export default function ProductFinderModal(props) {
 ProductFinderModal.propTypes = {
   modalIsOpen: PropTypes.bool.isRequired,
   setIsOpen: PropTypes.func.isRequired,
-  setSelectedProduct: PropTypes.func.isRequired,
+  selectedProduct: PropTypes.shape({
+    commodity_name: PropTypes.string,
+    commodity_code: PropTypes.string,
+  }),
+}
+ProductFinderModal.defaultProps = {
+  selectedProduct: null
 }